Fixed ball phasing bug
By moving the ball afterward instead of before, we were effectively just
moving the ball in a new direction based on the accelerometer instead of
reverting it's position. Switched back to old style to undo bug.
Another potential fix is to locally check coords, but would result in a
similar workload.